end to end delay and decimal timer

+++ src/Log.java
@@ -3,10 +3,10 @@ public class Log {
 	
 	public int taskId;
 	public LogType logType;
-	public int time;
+	public double time;
 	public LogSeverity logSeverity;
 	
-	public Log(int _taskId, int _time, LogType _logType, LogSeverity _logSeverity) {
+	public Log(int _taskId, double _time, LogType _logType, LogSeverity _logSeverity) {
 		this.taskId = _taskId;
 		this.logType = _logType;
 		this.time = _time;
